feat: add digital permission feature and normalize bulk import logic

+++ app/Database/Migrations/2023-08-18-000004_AddSuperadmin.php
@@ -25,6 +25,8 @@ class AddSuperadmin extends Migration
 
     public function down()
     {
-        $this->forge->dropColumn('users', 'is_superadmin');
+        if ($this->db->fieldExists('is_superadmin', 'users')) {
+            $this->forge->dropColumn('users', 'is_superadmin');
+        }
     }
 }
